requirements.txt
autopep8==1.5.7
awscrt==0.12.1
awsiotsdk==1.7.1
boto3==1.18.42
botocore==1.21.42
certifi==2021.5.30
charset-normalizer==2.0.5
idna==3.2
jmespath==0.10.0
pycodestyle==2.7.0
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
toml==0.10.2
urllib3==1.26.6
